+++ components/resultpage-components/ResultData.js
@@ -5,6 +5,7 @@ import Reviews from "./Reviews";
 import ReviewsMobile from "./ReviewsMobile";
 import SimilarBooks from "./SimilarBooks";
 import Meta from "../global-components/Meta";
+import ReviewBreakdown from "./ReviewBreakdown";
 
 const ResultData = ({ scrapedData }) => {
   const [imageLoaded, setImageLoaded] = useState(false);
@@ -191,7 +192,7 @@ const ResultData = ({ scrapedData }) => {
               </h2>
               <span className="text-md">{scrapedData.lang}</span>
             </div>
-            <div id="bookURL" className="mb-6">
+            <div id="bookURL" className="hidden sm:flex flex-col mb-4">
               <h2
                 className={
                   scrapedData.scrapeURL
@@ -202,7 +203,7 @@ const ResultData = ({ scrapedData }) => {
                 Goodreads URL:{" "}
               </h2>
               <a
-                className="text-blue-600 dark:text-blue-500 underline"
+                className="text-blue-600 dark:text-blue-500 underline truncate"
                 target="_blank"
                 rel="noreferrer"
                 href={
@@ -211,14 +212,17 @@ const ResultData = ({ scrapedData }) => {
                     : `https://www.goodreads.com/book/show/${scrapedData.scrapeURL}`
                 }
               >
-                <span className=" text-xs lg:text-md">
+                <p className="text-sm">
                   {scrapedData.scrapeURL.includes("https://www.goodreads.com")
                     ? scrapedData.scrapeURL
                     : `https://www.goodreads.com/book/show/${scrapedData.scrapeURL}`}
-                </span>
+                </p>
               </a>
             </div>
 
+            {scrapedData.reviewBreakdown && (
+              <ReviewBreakdown data={scrapedData} />
+            )}
             <div className="block lg:hidden">
               {scrapedData.related != "" && (
                 <SimilarBooks data={scrapedData.related} mobile={true} />