+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - Feb 10, 2023
+
+### Added
+
+- Add review breakdown
+
+### Changed
+
+- Don't show search result components when there is a scraper error
+- Dark mode review background color
+
+## [2.4.2] - Feb 5, 2023
+
+### Fixed
+
+- More screenshots link in the README not working on Codeberg
+
 ## [2.4.1] - Feb 5, 2023
 
 ### Added